requirements.txt
numpy==1.11.1
psutil==4.3.0
Django==1.8.10
djangorestframework==3.3.2
zeroc-ice==3.6.1
uwsgi==2.0.13.1
xlsxwriter==0.9.3
requests==2.9.1
pycrypto==2.6.1
taskflow==2.10.0
json_compare==0.2.11